+++ packages/extension-base/src/koni/background/subscription.ts
@@ -217,7 +217,7 @@ export class KoniSubscription {
   initCrowdloanSubscription (addresses: string[], substrateApiMap: Record<string, _SubstrateApi>, onlyRunOnFirstTime?: boolean) {
     const subscriptionPromise = subscribeCrowdloan(addresses, substrateApiMap, (networkKey, rs) => {
       this.state.setCrowdloanItem(networkKey, rs);
-    }, this.state.getChainInfoMap());
+    });
 
     if (onlyRunOnFirstTime) {
       subscriptionPromise.then((unsub) => {